requirements.txt
streamlit==0.81.0
google-cloud-texttospeech==2.3.0
ipython==7.23.0